# Async runtime with graph-based dependency cycle checks