--- a/hakyll.cabal
+++ b/hakyll.cabal
@@ -169,6 +169,7 @@ Library
     Paths_hakyll
 
   Build-Depends:
+    array                >= 0.5      && < 1,
     base                 >= 4.8      && < 5,
     binary               >= 0.5      && < 0.10,
     blaze-html           >= 0.5      && < 0.10,
@@ -181,6 +182,7 @@ Library
     file-embed           >= 0.0.10.1 && < 0.0.14,
     filepath             >= 1.0      && < 1.5,
     hashable             >= 1.0      && < 2,
+    lifted-async         >= 0.10     && < 1,
     lrucache             >= 1.1.1    && < 1.3,
     mtl                  >= 1        && < 2.3,
     network-uri          >= 2.6      && < 2.7,
@@ -191,6 +193,7 @@ Library
     regex-tdfa           >= 1.1      && < 1.4,
     resourcet            >= 1.1      && < 1.3,
     scientific           >= 0.3.4    && < 0.4,
+    stm                  >= 2.3      && < 3,
     tagsoup              >= 0.13.1   && < 0.15,
     template-haskell     >= 2.14     && < 2.17,
     text                 >= 0.11     && < 1.3,