--- a/hakyll.cabal
+++ b/hakyll.cabal
@@ -183,7 +183,6 @@ Library
     filepath             >= 1.0      && < 1.5,
     hashable             >= 1.0      && < 2,
     lifted-async         >= 0.10     && < 1,
-    lifted-base          >= 0.2      && < 1,
     lrucache             >= 1.1.1    && < 1.3,
     mtl                  >= 1        && < 2.3,
     network-uri          >= 2.6      && < 2.7,