--- a/lib/Hakyll/Core/Runtime.hs
+++ b/lib/Hakyll/Core/Runtime.hs
@@ -5,8 +5,11 @@ module Hakyll.Core.Runtime
 
 
 --------------------------------------------------------------------------------
-import           Control.Concurrent.Async.Lifted (forConcurrently_)
+import           Control.Concurrent              (getNumCapabilities)
+import           Control.Concurrent.Async.Lifted (mapConcurrently_)
+import           Control.Concurrent.QSemN        (QSemN, newQSemN, signalQSemN, waitQSemN)
 import           Control.Concurrent.STM          (atomically, modifyTVar', readTVarIO, newTVarIO, TVar)
+import           Control.Exception.Lifted        (bracket_)
 import           Control.Monad                   (unless)
 import           Control.Monad.Except            (ExceptT, runExceptT, throwError)
 import           Control.Monad.Reader            (ask)
@@ -61,6 +64,10 @@ run config logger rules = do
     mOldFacts <- Store.get store factsKey
     let (oldFacts) = case mOldFacts of Store.Found f -> f
                                        _             -> mempty
+    
+    -- Number of physical cores available for compilation
+    numproc <- getNumCapabilities 
+    Logger.debug logger $ "Using at most " ++ " cores..."
 
     -- Build runtime read/state
     let compilers = rulesCompilers ruleSet
@@ -71,6 +78,7 @@ run config logger rules = do
             , runtimeStore         = store
             , runtimeRoutes        = rulesRoutes ruleSet
             , runtimeUniverse      = M.fromList compilers
+            , runtimeCapabilities  = numproc
             }
 
     state <- newTVarIO $ RuntimeState 
@@ -110,6 +118,7 @@ data RuntimeRead = RuntimeRead
     , runtimeStore         :: Store
     , runtimeRoutes        :: Routes
     , runtimeUniverse      :: Map Identifier (Compiler SomeItem)
+    , runtimeCapabilities  :: Int
     }
 
 
@@ -186,11 +195,24 @@ scheduleOutOfDate = do
 --------------------------------------------------------------------------------
 pickAndChase :: Runtime ()
 pickAndChase = do
+    numproc <- runtimeCapabilities <$> ask
     todo <- runtimeTodo <$> getRuntimeState
     unless (null todo) $ do
         checkForDependencyCycle
-        forConcurrently_ (M.keys todo) chase
+        mapConcurrentlyN_ numproc chase (M.keys todo)
         pickAndChase
+
+
+--------------------------------------------------------------------------------
+-- | Maps a function and discard the results, performing at most @N@ actions concurrently.
+mapConcurrentlyN_ :: Traversable t => Int -> (a -> Runtime ()) -> t a -> Runtime ()
+mapConcurrentlyN_ n f xs = do
+  -- Emulating a pool of processes with locked access
+  sem <- liftIO $ newQSemN n
+  mapConcurrently_ (with sem . f) xs
+  where
+    with :: QSemN -> Runtime a -> Runtime a
+    with s = bracket_ (liftIO $ waitQSemN s 1) (liftIO $ signalQSemN s 1)
 
 
 --------------------------------------------------------------------------------