Hello,

This is a PR to make hakyll's runtime concurrent. 

The general idea is based on #725. That particular implementation incorrectly identified a dependency cycle [in my personal website](https://github.com/LaurentRDC/personal-website) that did not exist. Hence, this PR includes more robust dependency cycle checking based on `Data.Graph`. A test case has been added to ensure this works.

The concurrency is implemented as follows.
1. Identify all items that need to be compiled;
2. Compile all items concurrently. Items that compiled successfully are removed from the todo list, while those with dependencies are kept;
3. The remaining todo and their dependencies are checked for cycles in a single thread;
4. Go to step 1 until all items have been compiled.

```haskell
pickAndChase :: Runtime ()
pickAndChase = do
    todo <- fmap runtimeTodo . liftIO . readTVarIO =<< get
    unless (null todo) $ do
        checkForDependencyCycle
        forConcurrently_ (M.keys todo) chase
        pickAndChase
```

The advantage of this implementation is that it is easy to swap between concurrent/serial by simply swapping `forConcurrently_` for `forM_`. Maybe we can offer this option as a compile flag?

I have tested this version of hakyll on my personal page and I'm very satisfied with the results.

closes #714