@LaurentRDC 2021-04-23
Here is the benchmark on a site with quite a few JPGs to compress with [hakyll-images](https://github.com/LaurentRDC/hakyll-images):

| Pre-PR                  | 6148 ms |
|-------------------------|---------|
| This PR `+RTS -N1 -RTS` | 6179 ms |
| This PR `+RTS -N4 -RTS` | 4366 ms |