@Minoru 2021-04-23
This shows noticeable improvements [on my own site](https://github.com/Minoru/blog.debiania.in.ua):

```
$ hyperfine \
    --prepare './debiania-old clean' './debiania-old build' \
    --prepare './debiania-new clean' './debiania-new build +RTS -N4'

Benchmark #1: ./debiania-old build
  Time (mean ± σ):     15.345 s ±  0.153 s    [User: 40.101 s, System: 15.224 s]
  Range (min … max):   15.168 s … 15.628 s    10 runs

Benchmark #2: ./debiania-new build +RTS -N4
  Time (mean ± σ):     10.556 s ±  0.219 s    [User: 22.514 s, System: 4.638 s]
  Range (min … max):   10.096 s … 10.848 s    10 runs
```

There is some scalability weirdness though. I have Intel Core i7 7700HQ, which is a mobile quad-core with HyperThreading. Theoretically, 8-thread workload should slightly outperform 4-thread, but it's not the case:

```
$ hyperfine \
    --prepare './debiania-new clean' './debiania-new build +RTS -N4' \
    --prepare './debiania-new clean' './debiania-new build +RTS -N8' \
    --prepare './debiania-new clean' './debiania-new build +RTS -N'

Benchmark #1: ./debiania-new build +RTS -N4
  Time (mean ± σ):     10.556 s ±  0.219 s    [User: 22.514 s, System: 4.638 s]
  Range (min … max):   10.096 s … 10.848 s    10 runs

Benchmark #2: ./debiania-new build +RTS -N8
  Time (mean ± σ):     12.546 s ±  0.224 s    [User: 48.112 s, System: 12.646 s]
  Range (min … max):   12.122 s … 12.871 s    10 runs

Benchmark #3: ./debiania-new build +RTS -N
  Time (mean ± σ):     12.766 s ±  0.243 s    [User: 48.934 s, System: 13.672 s]
  Range (min … max):   12.200 s … 13.071 s    10 runs
```

It's still faster than a single-thread, so not a blocker for this PR; just curious if anyone has any insight into what might be causing this.

(The numbers in two benchmarks match exactly because it's actually a single benchmark whose report I split in two.)