@LaurentRDC 2021-04-23
I had this problem with [pandoc-plot](github.com/LaurentRDC/pandoc-plot) where if the number of async jobs was unlimited, resource contention would slow down the overall running time.

Let me try to limit the number of concurrent jobs to the number of cores and see if scaling makes more sense.