@Minoru 2021-04-23
@LaurentRDC Thanks for looking into that! I thought it might be related to my [`gzipFileCompiler`](https://github.com/Minoru/blog.debiania.in.ua/blob/79d3f718a2f6214b1781fcd33543b15bcdc71a7b/src/Debiania/Compilers.hs#L56-L61), which compresses every page I have, but removing it didn't fix the scalability issue.