@LaurentRDC 2021-04-23
Indeed, limiting the number of concurrent tasks helps with scaling. After the most recent patch, building my site with 1 - 4 cores yields the following results:

|     	|      	|
|-----	|------	|
| -N1 	| 153s 	|
| -N2 	| 80s  	|
| -N3 	| 63s  	|
| -N4 	| 55s  	|